requirements.txt
langchain
langchain_community
pandasai
ollama
streamlit
openai